Improve performance of fixer
Several optimizations were tested, and the most effective appears to
be to avoid parent node lookups when possible. The node counting which
was being used to do this same work was found to offer only marginal
benefit in performance testing, and has therefore been removed, since
it trades one cost for another -- it requires more invocations of our
node visitor functions.