Add optimistic result caching
There is no caching here for failing files. For now, only successful
(passing) files are cached, in a naive filesystem-as-a-key-value-store
cache directory of hashes.

This greatly improves `slyp` performance when files are unchanged and
the repo being checked fully passes, by avoiding doing all of the work
of running linters and fixers altogether.

In order to support future changes to fixer behavior, and the
necessary cache invalidations, a "contract version" has been encoded
into the cache directories in use, `_v1_` to indicate that this is the
first `slyp` version.

This version is kept "bare" in the cache directory name so that it
will be possible to clean caches later.