Add tests and fix bug in AST branch matcher
When visiting an `If` node, `generic_visit` was being called on all
subnodes in the non-error case. This ends up skipping the `test` node
and the top node for each statement in the `body`. The correct logic
is to `generic_visit` any `If` nodes in the `orelse` branch (to avoid
re-comparing branches in an `if elif else` structure), but `visit` all
other subnodes.